requirements.txt
fastapi
uvicorn[standard]
aiosqlite
httpx
pydantic
python-dotenv
gunicorn
websockets